+++ product-service/src/main/java/com/ecommerce/product_service/dto/response/ProductResponse.java
@@ -2,6 +2,7 @@ package com.ecommerce.product_service.dto.response;
 
 import lombok.Builder;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 @Builder
@@ -9,8 +10,8 @@ public class ProductResponse {
     private CategoryResponse categoryResponse;
     private String name;
     private String description;
-    private double price;
-    private int stock;
+    private BigDecimal price;
+    private Integer stock;
     private String imageUrl;
     private boolean active;
     private LocalDateTime createdAt;
